Update README.md to document new Join extension method
The `README.md` file has been updated to include documentation for a new extension method called `Join`. This method allows for the concatenation of items in an enumerable into a single string, using a specified separator. This addition provides a convenient way to create a single string from a collection of items, enhancing the library's functionality.